requirements.txt
pandas>=1.3.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
mlxtend>=0.19.0
statsmodels>=0.13.0
pmdarima>=2.0.0
pyarrow>=10.0.0
pyyaml>=6.0
papermill>=2.4.0
jupyter>=1.0.0
tqdm>=4.62.0
streamlit>=1.20.0
